Make zshenv a bit more generic for path & manpath
The goal here is to just specify base-paths, and have zshenv add the
usual suspects onto the base-path, and if they exist as directories, add
them to $path and $manpath. this way, the paths are simpler to search at
runtime. more importantly, when I make a change to the paths, the man
path will get added if it exists, too.

I'm also adding ./.local to the path so that each directory can have
customized binaries that act in localized ways. $HOME/.local continues
to be in the path no matter what.