feat(systems): update NixOS release to 24.11
This commit updates the NixOS release version to 24.11 for the `chickenpi`,
`dunbar`, and `wildwood` systems. This ensures that the systems are compatible
with the latest NixOS release and allows for access to the latest packages and
features.

Additionally, the commit removes the deprecated `sound.enable` configuration
from the `dunbar` and `wildwood` systems, as sound is now configured via
`pipewire`.

The commit also adds the `isNormalUser` and `home` configuration for the `clord`
and `eugene` users in the `flake.nix` file, ensuring that their user accounts
are properly set up.